pyproject.toml
[project]
name = "scoutra-research-agent"
version = "0.1.0"
description = "AI Research Agent for Scoutra AI — automated business opportunity analysis"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.43.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
]

[project.scripts]
scoutra-agent = "research_agent.agent:main"
